requirements.txt
opencv-python
opencv-contrib-python
numpy
pandas
Pillow
mediapipe
tk
